[feature/SBOT-215]: 정렬 선택 후 상태를 Back Result Api로 전달하도록 변경
